+++ src/pyside_ext/elements/base.py
@@ -33,7 +33,8 @@ class BasePanelElement:
         self.handler = handler
 
     @abstractmethod
-    def setup(self): ...
+    def setup(self):
+        pass
 
     def configure(self, *args, **kwargs):
         raise NotImplementedError(f"Configure method not implemented for {self.__class__.__name__}.")